requirements.txt
tqdm
PyYAML
gradio
numpy
pydub
scipy